app/requirements.txt
django==3.0.6
psycopg2-binary==2.8.3
djangorestframework==3.11.0
django-filter==2.2.0
markdown
Pillow==7.1.2
djoser==2.1.0
django-cors-headers
sentry-sdk